// src/pinecone-generated-ts-fetch/db_control/models/PatchIndexDeploymentRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Pinecone Control Plane API
 * Pinecone is a vector database that makes it easy to search and retrieve billions of high-dimensional vectors.
 *
 * The version of the OpenAPI document: 2026-07
 * Contact: support@pinecone.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * Deployment update parameters for pod-based indexes. Specify `replicas`, `pod_type`, or both. Deployment type, cloud/region, and environment cannot be changed after creation. Do not include a `deployment_type` field; it is not accepted.
 * @export
 * @interface PatchIndexDeploymentRequest
 */
export interface PatchIndexDeploymentRequest {
    /**
     * The number of replicas. Replicas duplicate your index. They provide higher availability and throughput. Replicas can be scaled up or down as your needs change.
     * @type {number}
     * @memberof PatchIndexDeploymentRequest
     */
    replicas?: number;
    /**
     * The type of pod to use. One of `s1`, `p1`, or `p2` appended with `.` and one of `x1`, `x2`, `x4`, or `x8`.
     * @type {string}
     * @memberof PatchIndexDeploymentRequest
     */
    podType?: string;
}

/**
 * Check if a given object implements the PatchIndexDeploymentRequest interface.
 */
export function instanceOfPatchIndexDeploymentRequest(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function PatchIndexDeploymentRequestFromJSON(json: any): PatchIndexDeploymentRequest {
    return PatchIndexDeploymentRequestFromJSONTyped(json, false);
}

export function PatchIndexDeploymentRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): PatchIndexDeploymentRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'replicas': !exists(json, 'replicas') ? undefined : json['replicas'],
        'podType': !exists(json, 'pod_type') ? undefined : json['pod_type'],
    };
}

export function PatchIndexDeploymentRequestToJSON(value?: PatchIndexDeploymentRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'replicas': value.replicas,
        'pod_type': value.podType,
    };
}